+++ src/components/DatePicker/DatePicker.test.tsx
@@ -105,7 +105,10 @@ describe("DatePicker", () => {
       );
 
       expect(input).toBeInTheDocument();
-      expect(input).not.toHaveAttribute("aria-label");
+      expect(input).toHaveAttribute(
+        "aria-label",
+        "Select the full date you want to visit NYPL, Press tab to access the calendar."
+      );
       // Date format based on component specification yyyy-mm-dd.
       expect(date).toEqual(`${year}-${month}-${day}`);
       expect(screen.getByDisplayValue(date)).toBeInTheDocument();
@@ -596,20 +599,24 @@ describe("DatePicker", () => {
       // Helper text for the "To" input
       expect(screen.getByText(/Note for the 'to' field./i)).toBeInTheDocument();
 
-      const fromInput = screen.getByRole("textbox", { name: "From" });
-      const toInput = screen.getByRole("textbox", { name: "To" });
+      const inputFrom = screen.getByRole("textbox", {
+        name: "From, Press tab to access the calendar.",
+      });
+      const inputTo = screen.getByRole("textbox", {
+        name: "To, Press tab to access the calendar.",
+      });
 
       // The `fromInput` should have an `aria-describedby` value of both the id of
       // the `helperText` and the id of the `helperTextFrom` in that order - from
       // more general to more specific.
-      expect(fromInput).toHaveAttribute(
+      expect(inputFrom).toHaveAttribute(
         "aria-describedby",
         "datePicker-helper-text datePicker-start-helperText"
       );
       // The `toInput` should have an `aria-describedby` value of both the id of
       // the `helperText` and the id of the `helperTextTo` in that order - from
       // more general to more specific.
-      expect(toInput).toHaveAttribute(
+      expect(inputTo).toHaveAttribute(
         "aria-describedby",
         "datePicker-helper-text datePicker-end-helperText"
       );
@@ -645,8 +652,14 @@ describe("DatePicker", () => {
         />
       );
       // Both input fields are disabled.
-      expect(screen.getByLabelText(/From/i)).toHaveAttribute("disabled");
-      expect(screen.getByLabelText(/To/i)).toHaveAttribute("disabled");
+      let inputFrom = screen.getByRole("textbox", {
+        name: "From, Press tab to access the calendar.",
+      });
+      let inputTo = screen.getByRole("textbox", {
+        name: "To, Press tab to access the calendar.",
+      });
+      expect(inputFrom).toHaveAttribute("disabled");
+      expect(inputTo).toHaveAttribute("disabled");
 
       rerender(
         <DatePicker
@@ -659,11 +672,17 @@ describe("DatePicker", () => {
           isDateRange
         />
       );
-      // Both input fields are required.
+
+      inputFrom = screen.getByRole("textbox", {
+        name: "From, Press tab to access the calendar.",
+      });
+      inputTo = screen.getByRole("textbox", {
+        name: "To, Press tab to access the calendar.",
+      }); // Both input fields are required.
       // The "Required" text is only displayed once in the `legend`.
       expect(screen.getAllByText(/Required/i)).toHaveLength(1);
-      expect(screen.getByLabelText(/From/i)).toHaveAttribute("required");
-      expect(screen.getByLabelText(/To/i)).toHaveAttribute("required");
+      expect(inputFrom).toHaveAttribute("required");
+      expect(inputTo).toHaveAttribute("required");
     });
 
     // Note: Have to add initial dates so that the snapshot tests always
@@ -750,15 +769,19 @@ describe("DatePicker", () => {
           labelText="Select the date range you want to visit NYPL"
         />
       );
-      const fromInput = screen.getByLabelText(/From/i);
-      const toInput = screen.getByLabelText(/To/i);
+      const inputFrom = screen.getByRole("textbox", {
+        name: "From, Press tab to access the calendar.",
+      });
+      const inputTo = screen.getByRole("textbox", {
+        name: "To, Press tab to access the calendar.",
+      });
 
-      expect(fromInput).toHaveValue("1988-03-02");
-      expect(toInput).toHaveValue("1988-03-28");
+      expect(inputFrom).toHaveValue("1988-03-02");
+      expect(inputTo).toHaveValue("1988-03-28");
       // expect(screen.getAllByDisplayValue(date)).toHaveLength(2);
 
       // Let's select a new day.
-      userEvent.click(fromInput);
+      userEvent.click(inputFrom);
       // The popup displays. Select a new day.
       const newDateFrom = 5;
       const newDateTo = 25;
@@ -771,7 +794,7 @@ describe("DatePicker", () => {
       // expect(screen.getAllByDisplayValue(date)).toHaveLength(1);
 
       // Now select the "To" date.
-      userEvent.click(toInput);
+      userEvent.click(inputTo);
       // The popup displays.
       userEvent.click(screen.getByText(newDateTo));